requirements.txt
# ========================================
# Virtual Desk Assistant - Unified Requirements
# Python 3.10+ required
# ========================================

# ========================================
# Web Framework
# ========================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ========================================
# Database & ORM
# ========================================
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# ========================================
# Authentication & Security
# ========================================
python-jose[cryptography]==3.3.0
authlib==1.3.2

# ========================================
# HTTP Client
# ========================================
httpx==0.28.1

# ========================================
# Configuration
# ========================================
python-dotenv==1.0.1
pydantic==2.12.4
pydantic-settings>=2.10.1,<3.0.0
email-validator==2.2.0

# ========================================
# OpenAI & LLM
# ========================================
openai==1.109.1

# ========================================
# LangChain & RAG (보고서 Agent 시스템 포함)
# ========================================
langchain==1.1.0
langchain-core==1.1.0
langchain-community==0.4
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
ormsgpack==1.12.0
xxhash==3.6.0
langsmith==0.4.43

# ========================================
# Vector Database & Search
# ========================================
chromadb==0.5.23
rank-bm25==0.2.2
#ddgs==7.5.1

# ========================================
# PDF Processing
# ========================================
PyMuPDF==1.24.13
pdfplumber==0.11.8
Pillow==11.0.0
pdfkit==1.2.0

# ========================================
# Utilities
# ========================================
python-dateutil==2.9.0.post0
rich==13.9.4
tqdm==4.66.5
tiktoken==0.8.0
numpy==2.1.3
pandas==2.2.3
openpyxl==3.1.5
tenacity==9.0.0
matplotlib==3.9.2
seaborn==0.13.2
# ========================================
# Testing
# ========================================
pytest==8.3.4
pytest-asyncio==0.24.0

# Google API 공식 SDK
google-api-python-client==2.147.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1



# Notion 공식 SDK
notion-client==2.2.1

# ========================================
# HTML Template Engine (보고서 HTML 생성)
# ========================================
jinja2==3.1.4

# ========================================
# Embedding Models (보고서 RAG)
# ========================================
sentence-transformers==3.3.1

# ========================================
# Notes
# ========================================
# 통합 버전 (2025-12-03)
# - chromadb==0.5.23: 브레인스토밍 데이터 호환성
# - pydantic==2.12.4: 최신 안정 버전
# - langchain==1.1.0: LangChain 1.x 버전 업그레이드 (보고서 Agent 시스템)
# - langgraph==1.0.4: LangGraph 지원 추가
# - langchain-community>=0.3.0: 보고서 Agent 호환
# - langchain-openai>=0.2.0: 보고서 Agent 호환
# - pydantic-settings>=2.10.1: 보고서 Agent 호환
# - jinja2==3.1.4: HTML 보고서 생성 (추가됨)
# - sentence-transformers==3.3.1: HuggingFace 임베딩 모델 (보고서 RAG용 추가됨)
#
# 제거된 패키지 (확인된 미사용):
# - torch: 미사용
# - redis: 설정만 있고 미사용

# ========================================
# NPM Dependencies (Node.js)
# ========================================
# npm 패키지는 package.json에서 관리됩니다.
#
# ⚠️ npm이 설치되지 않은 경우:
# 1. Node.js 설치 (npm 포함): https://nodejs.org/
# 2. LTS 버전 다운로드 및 설치
# 3. 설치 확인:
#    node -v
#    npm -v
# 4. 설치 후: npm install
#
# 자세한 설치 방법은 WINDOWS_SETUP.md 참조
#
# DevDependencies:
# - concurrently: ^8.2.2 (동시 실행)
# - electron: ^39.1.2 (Electron 프레임워크)
# - wait-on: ^8.0.0 (대기 유틸리티)

